PROD-677 Fiddling with things to get the tests to fail like they do in prod
